refactor(nms): Separate the GatewayContext
